+++ src/app/modules/search-parameters/search-parameters.module.ts
@@ -1,18 +1,18 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SearchParametersComponent } from './search-parameters.component';
-import { SearchParameterModule } from '../search-parameter/search-parameter.module';
 import { MatIconModule } from '@angular/material/icon';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 import { MatFormFieldModule } from '@angular/material/form-field';
+import { SearchParameterGroupModule } from '../search-parameter-group/search-parameter-group.module';
 
 @NgModule({
   declarations: [SearchParametersComponent],
   exports: [SearchParametersComponent],
   imports: [
     CommonModule,
-    SearchParameterModule,
+    SearchParameterGroupModule,
     MatIconModule,
     ReactiveFormsModule,
     MatButtonModule,