+++ src/app/modules/stepper/stepper.component.ts
@@ -89,7 +89,7 @@ export class StepperComponent implements OnDestroy {
       serviceBaseUrl: this.fhirBackend.serviceBaseUrl,
       maxPatientCount: this.defineCohortComponent.defineCohortForm.value
         .maxPatientsNumber,
-      rawCriteria: this.defineCohortComponent.patientParams.parameterList.value,
+      rawCriteria: this.defineCohortComponent.patientParams.parameterGroupList.getRawValue(),
       data:
         this.viewCohortComponent?.resourceTableComponent?.dataSource?.data ??
         [],
@@ -131,9 +131,9 @@ export class StepperComponent implements OnDestroy {
             .get('maxPatientsNumber')
             .setValue(maxPatientCount);
           // Set search parameter form values.
-          this.defineCohortComponent.patientParams.parameterList.clear();
+          this.defineCohortComponent.patientParams.parameterGroupList.clear();
           (rawCriteria as SearchParameter[]).forEach((searchParam) => {
-            this.defineCohortComponent.patientParams.parameterList.push(
+            this.defineCohortComponent.patientParams.parameterGroupList.push(
               new FormControl(searchParam)
             );
           });